[chore][pkg/stanza] Cleanup windows input operator files
