+++ pkg/stanza/operator/input/windows/config_all.go
